chore(tcp): move tcp processing to own file
